users-service/requirements.txt
fastapi
uvicorn
python-multipart
factory-boy
SQLAlchemy
sqlmodel
pydantic[email]
aiosqlite
alembic
passlib
bcrypt
python-jose[cryptography]
pytest
pytest-asyncio
anyio
httpx
requests
trio